Add support for max client limit for pending objects in outbound queue
This is done with a server level config option, which is disabled by default.
It is also changeable at runtime with the CONFIG SET command.

    New config parameter:  maxclientqueue
    Default: 0 (no limit)

The implementation checks in two places to see if a client has overflowed,
at the time of appending more data to a client's queue, and also on a 30 sec
timer, to clean up overflowed clients that are now idle.

Added by J. Will Pierce willp@nuclei.com